pyproject.toml
[project]
name = "mstemplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.8.2",
    "fastapi[standard]>=0.112.4",
    "ms-core",
    "pydantic>=2.11.7",
    "tortoise-orm[asyncpg]>=0.21.7",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "redis>=7.2.0",
    "loguru>=0.7.3",
    "google-auth>=2.48.0",
]
[tool.pyright]
reportInvalidTypeForm = "none"

[tool.uv.sources]
ms-core = { url = "https://github.com/HexChap/MSCore/archive/refs/heads/master.zip" }

[tool.aerich]
tortoise_orm = "main.tortoise_conf"
location = "./migrations"
src_folder = "./."
